HLE: Move NS:S into APT and remove NS
